streaming/requirements.txt
pyspark==3.1.2
python-dotenv
cassandra-driver